Added Delimiter Support
Delimiters have now been added in to the AST.  This is because there
was a bug where delimiters were stripped from IF statements.  The
EQ operator of an IF is '==', while '=' is also a valid delimiter,
meaning IF statements went from:

  if 1 == 1 echo test

to:

  if 1 1 echo test

This is now fixed, along with the test suite.  Also added a ROP-style
return from the AST parser.  Still need to add tests for checking
parser-failure events.