Started moving IF, REM, and FOR identification to parse-phase
Attempted to identify and handle FOR loops in the interpreter
which was causing a lot of headaches around delayed expansion
expanding valid FOR-loop statements.  Research shows that IF,
FOR, and REM statements are identified much earlier at parse-time.
As a result, have started modifiying the parser so that the
AST it produces contains five statement types:

  1. Operators
  2. FOR loops
  3. IF statements
  4. REM statements
  5. Command statements