Added specific types for <0, 0, and >0 ints, and fixed failing tests
We no longer return a 'Number' type for ranges.  Instead, we return
either a 'NegativeInt', 'Zero', or 'PositiveInt'.  This means we can
perform better (and cleaner) error checking.

Added some REALLY HORRIBLE code for handling and detecting failing
tests.  I am sure there's a better way -- so in the future, when
I've leared a bit more about F#, I'll revisit and fix this mess, but
for now, it'll have to do.