Added support for 'Empty' values
There are lots of edge cases with how the 'FOR /f ...' parser works.
This commit fixes one-such edge case where the following EXPR is
not valid:

  FOR /f "tokens= eol=," ...

however, the following is valid:

  FOR /f "tokens= " ...

New code handles the case where the value of the 'tokens=' keyword
is match-able as an 'empty' value, so the parser can check whether
it's allowed to be empty, given the current expr string.